Changed callback, added color tags
Changed the callback function so it displays the most similar sentence on the HTML and added "color" tags that we can change and play around with in the future. The idea is we can change the tags from colors to "feelings" like sad, angry, happy and so on and add sounds/speech that we record ourselves accordingly.